backEnd/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-docx